fix(controller): Add EnvironmentFile to the serivce files created by fleet.py
The service file that is created by the controller after a succesfull
build of a new container, didn't have access to the /etc/environment file
in CoreOS.

People that run etcd on the non-standard 127.0.0.1:4001 address, need to
make the etcd address availalbe to fleetctl in the .service file.
If you don't this, the etcdctl commands used in the ExecStartPre won't
be able to communicate with the correct etcd service.